+++ core/cli_entry.py
@@ -14,23 +14,16 @@ EXIT_CODE_ERROR = 1
 EXIT_CODE_SUCCESS = 0
 
 
-def validate_and_upload_single_entry(directory, force_upload=False):
+def upload_run_single_entry(directory, force_upload=False):
     """
     This function acts as a single point of entry for uploading a directory
 
-    Handles parsing and validating the directory for samples
-    Sets up the api layer based on config file
-    Verifies samples is able to be uploaded (verifies projects exist)
-    Initializes objects/routes on IRIDA to accept Samples (creates samples if they don't exist)
-    Starts the upload
+    Handles getting a directories run status, and running if conditions are met (valid run, new run or forced upload).
 
     :param directory: Directory of the sequencing run to upload
     :param force_upload: When set to true, the upload status file will be ignored and file will attempt to be uploaded
     :return:
     """
-    logging_start_block(directory)
-    logging.debug("validate_and_upload_single_entry:Starting {}".format(directory))
-
     directory_status = parsing_handler.get_run_status(directory)
     # Check if a run is invalid, an invalid run cannot be uploaded.
     if directory_status.status_equals(DirectoryStatus.INVALID):
@@ -48,10 +41,78 @@ def validate_and_upload_single_entry(directory, force_upload=False):
                           "You can bypass this error by uploading with the --force argument.".format(directory))
             return exit_error()
 
+    return _validate_and_upload(directory_status)
+
+
+def batch_upload_single_entry(batch_directory, force_upload=False):
+    """
+    This function acts as a single point of entry for batch uploading run directories
+
+    It uses _validate_and_upload as it function for uploading the individual runs
+
+    A list of runs to be uploaded is generated at start up, and all found runs will be attempted to be uploaded.
+
+    :param batch_directory: Directory containing sequencing run directories to upload
+    :param force_upload: When set to true, the upload status file will be ignored and file will attempt to be uploaded
+    :return:
+    """
+    logging.debug("batch_upload_single_entry:Starting {} with force={}".format(batch_directory, force_upload))
+    # get all potential directories to upload
+    directory_status_list = parsing_handler.get_run_status_list(batch_directory)
+    # list info about directories found
+    logging.info("Found {} potential run directories".format(len(directory_status_list)))
+    for directory_status in directory_status_list:
+        logging.info("DIRECTORY: %s\n"
+                     "%30sSTATUS:  %s\n"
+                     "%30sDETAILS: %s"
+                     % (directory_status.directory, "", directory_status.status, "", directory_status.message))
+
+    # if `force` is on, only don't upload invalid runs
+    if force_upload:
+        upload_list = [x for x in directory_status_list if not x.status_equals(DirectoryStatus.INVALID)]
+        logging.info("Starting upload for all non invalid runs. {} run(s) found. "
+                     "(Running with --force)".format(len(upload_list)))
+    # without `force` only upload new runs
+    else:
+        upload_list = [x for x in directory_status_list if x.status_equals(DirectoryStatus.NEW)]
+        logging.info("Starting upload for all new runs. {} run(s) found.".format(len(upload_list)))
+
+    # run upload, keep track of which directories did not upload
+    error_list = []
+    for directory_status in upload_list:
+        logging.info("Starting upload for {}".format(directory_status.directory))
+        result = _validate_and_upload(directory_status)
+        if result == EXIT_CODE_ERROR:
+            error_list.append(directory_status.directory)
+
+    logging.info("Uploads completed with {} error(s)".format(len(error_list)))
+    for directory in error_list:
+        logging.warning("Directory '{}' upload exited with ERROR, check log and status file for details"
+                        "".format(directory))
+
+    logging.info("Batch upload complete, Exiting!")
+    return exit_success()
+
+
+def _validate_and_upload(directory_status):
+    """
+    This function attempts to upload a single run directory
+
+    Handles parsing and validating the directory for samples
+    Sets up the api layer based on config file
+    Verifies samples is able to be uploaded (verifies projects exist)
+    Initializes objects/routes on IRIDA to accept Samples (creates samples if they don't exist)
+    Starts the upload
+
+    :param directory_status: DirectoryStatus object that has directory to try upload
+    :return:
+    """
+    logging_start_block(directory_status.directory)
+    logging.debug("upload_run_single_entry:Starting {}".format(directory_status.directory))
+
     # Add progress file to directory
     try:
-        directory_status.status = DirectoryStatus.PARTIAL
-        progress.write_directory_status(directory_status)
+        _set_and_write_directory_status(directory_status, DirectoryStatus.PARTIAL)
     except progress.exceptions.DirectoryError as e:
         logging.error("ERROR! Error while trying to write status file to directory {} with error message: {}"
                       "".format(e.directory, e.message))
@@ -60,21 +121,23 @@ def validate_and_upload_single_entry(directory, force_upload=False):
 
     # Do parsing (Also offline validation)
     try:
-        sequencing_run = parsing_handler.parse_and_validate(directory)
+        sequencing_run = parsing_handler.parse_and_validate(directory_status.directory)
     except parsers.exceptions.DirectoryError as e:
         # Directory was not valid for some reason
-        logging.error("ERROR! An error occurred with directory '{}', with message: {}".format(e.directory, e.message))
+        full_error = "ERROR! An error occurred with directory '{}', with message: {}".format(e.directory, e.message)
+        logging.error(full_error)
         logging.info("Samples not uploaded!")
-        directory_status.status = DirectoryStatus.ERROR
-        progress.write_directory_status(directory_status)
+        _set_and_write_directory_status(directory_status, DirectoryStatus.ERROR, full_error)
         return exit_error()
     except parsers.exceptions.ValidationError as e:
         # Sequencing Run / SampleSheet was not valid for some reason
-        logging.error("ERROR! Errors occurred during validation with message: {}".format(e.message))
-        logging.error("Error list: " + pformat(e.validation_result.error_list))
+        error_msg = "ERROR! Errors occurred during validation with message: {}".format(e.message)
+        logging.error(error_msg)
+        error_list_msg = "Error list: " + pformat(e.validation_result.error_list)
+        logging.error(error_list_msg)
         logging.info("Samples not uploaded!")
-        directory_status.status = DirectoryStatus.ERROR
-        progress.write_directory_status(directory_status)
+        full_error = error_msg + ", " + error_list_msg
+        _set_and_write_directory_status(directory_status, DirectoryStatus.ERROR, full_error)
         return exit_error()
 
     # Initialize the api for first use
@@ -85,8 +148,8 @@ def validate_and_upload_single_entry(directory, force_upload=False):
         logging.error("ERROR! Could not initialize irida api.")
         logging.error("Errors: " + pformat(e.args))
         logging.info("Samples not uploaded!")
-        directory_status.status = DirectoryStatus.ERROR
-        progress.write_directory_status(directory_status)
+        full_error = "ERROR! Could not initialize irida api. Errors: " + pformat(e.args)
+        _set_and_write_directory_status(directory_status, DirectoryStatus.ERROR, full_error)
         return exit_error()
     logging.info("*** Connected ***")
 
@@ -96,8 +159,8 @@ def validate_and_upload_single_entry(directory, force_upload=False):
     except api.exceptions.IridaConnectionError as e:
         logging.error("Lost connection to Irida")
         logging.error("Errors: " + pformat(e.args))
-        directory_status.status = DirectoryStatus.ERROR
-        progress.write_directory_status(directory_status)
+        full_error = "Lost connection to Irida. Errors: " + pformat(e.args)
+        _set_and_write_directory_status(directory_status, DirectoryStatus.ERROR, full_error)
         return exit_error()
 
     if not validation_result.is_valid():
@@ -105,8 +168,8 @@ def validate_and_upload_single_entry(directory, force_upload=False):
         logging.error("Sequencing run can not be uploaded. Encountered {} errors"
                       "".format(validation_result.error_count()))
         logging.error("Errors: " + pformat(validation_result.error_list))
-        directory_status.status = DirectoryStatus.ERROR
-        progress.write_directory_status(directory_status)
+        full_error = "Sequencing run can not be uploaded, Errors: " + pformat(validation_result.error_list)
+        _set_and_write_directory_status(directory_status, DirectoryStatus.ERROR, full_error)
         return exit_error()
     logging.info("*** Run Verified ***")
 
@@ -117,26 +180,43 @@ def validate_and_upload_single_entry(directory, force_upload=False):
     except api.exceptions.IridaConnectionError as e:
         logging.error("Lost connection to Irida")
         logging.error("Errors: " + pformat(e.args))
-        directory_status.status = DirectoryStatus.ERROR
-        progress.write_directory_status(directory_status)
+        full_error = "Lost connection to Irida. Errors: " + pformat(e.args)
+        _set_and_write_directory_status(directory_status, DirectoryStatus.ERROR, full_error)
         return exit_error()
     logging.info("*** Upload Complete ***")
 
     # Set progress file to complete
     try:
-        directory_status.status = DirectoryStatus.COMPLETE
-        progress.write_directory_status(directory_status, run_id=run_id)
+        _set_and_write_directory_status(directory_status, DirectoryStatus.COMPLETE, run_id=run_id)
     except progress.exceptions.DirectoryError as e:
         # this is an exceptionally rare case (successful upload, but fails to write progress)
         logging.ERROR("ERROR! Error while trying to write status file to directory {} with error message: {}"
                       "".format(e.directory, e.message))
         logging.info("Samples were uploaded, but progress file may be incorrect!")
 
-    logging.info("Samples in directory '{}' have finished uploading!".format(directory))
+    logging.info("Samples in directory '{}' have finished uploading!".format(directory_status.directory))
 
     logging_end_block()
 
     return exit_success()
+
+
+def _set_and_write_directory_status(directory_status, status, message=None, run_id=None):
+    """
+    Given a DirectoryStatus object, sets the status and message, and then writes to the directory status directory
+
+    :param directory_status: DirectoryStatus object
+    :param status: a valid DirectoryStatus status
+    :param message: string
+    :param run_id: optional, if provided, the run id and irida instance will be included when written
+    :return:
+    """
+    directory_status.status = status
+    directory_status.message = message
+    if run_id:
+        progress.write_directory_status(directory_status, run_id)
+    else:
+        progress.write_directory_status(directory_status)
 
 
 def exit_error():
@@ -179,4 +259,3 @@ def logging_end_block():
     logging.info("----------------ENDING UPLOAD RUN-----------------")
     logging.info("==================================================")
     logger.remove_directory_logger()
-