+++ parsers/__init__.py
@@ -1,2 +1,2 @@
-from .parsers import Parser
+from .parsers import Parser, supported_parsers
 from . import exceptions