+++ tests/core/test_cli_entry.py
@@ -1,5 +1,5 @@
 import unittest
-from unittest.mock import patch, MagicMock, Mock
+from unittest.mock import patch, MagicMock, Mock, call
 from os import path
 import os
 
@@ -12,9 +12,10 @@ if len(path_to_module) == 0:
     path_to_module = '.'
 
 
-class TestValidateAndUploadSingleEntry(unittest.TestCase):
+class TestUploadRunSingleEntry(unittest.TestCase):
     """
-    Tests the core.cli_entry.validate_and_upload_single_entry function
+    Tests the core.cli_entry.upload_run_single_entry function
+    Indirectly does testing on core.cli_entry._validate_and_upload function too
     """
 
     def setUp(self):
@@ -42,13 +43,15 @@ class TestValidateAndUploadSingleEntry(unittest.TestCase):
         Makes sure that all functions are called when a valid directory in given
         :return:
         """
+        my_directory = path.join(path_to_module, "fake_ngs_data")
+
         class StubValidationResult:
             @staticmethod
             def is_valid():
                 return True
 
         class StubDirectoryStatus:
-            directory = ""
+            directory = my_directory
             status = DirectoryStatus.NEW
 
             @staticmethod
@@ -56,21 +59,20 @@ class TestValidateAndUploadSingleEntry(unittest.TestCase):
                 return status == DirectoryStatus.NEW
 
         # add mock data to the function calls that are essential
-        mock_parsing_handler.get_run_status.side_effect = [StubDirectoryStatus]
+        stub_directory_status = StubDirectoryStatus()
+        mock_parsing_handler.get_run_status.side_effect = [stub_directory_status]
         mock_parsing_handler.parse_and_validate.side_effect = ["Fake Sequencing Run"]
         mock_api_handler.initialize_api_from_config.side_effect = [None]
         mock_api_handler.prepare_and_validate_for_upload.side_effect = [StubValidationResult]
         mock_api_handler.upload_sequencing_run.side_effect = [None]
         mock_progress.write_directory_status.side_effect = [None, None]
 
-        directory = path.join(path_to_module, "fake_ngs_data")
-
-        cli_entry.validate_and_upload_single_entry(directory, force_upload=False)
+        cli_entry.upload_run_single_entry(my_directory, force_upload=False)
 
         # Make sure directory status is init
-        mock_progress.write_directory_status.assert_called_with(StubDirectoryStatus, run_id=None)
+        mock_progress.write_directory_status.assert_called_with(stub_directory_status)
         # Make sure parsing and validation is done
-        mock_parsing_handler.parse_and_validate.assert_called_with(directory)
+        mock_parsing_handler.parse_and_validate.assert_called_with(my_directory)
         # api must be initialized
         mock_api_handler.initialize_api_from_config.assert_called_with()
         # api must prep for upload
@@ -90,12 +92,12 @@ class TestValidateAndUploadSingleEntry(unittest.TestCase):
         """
 
         directory = path.join(path_to_module, "fake_ngs_data")
+        directory_status = DirectoryStatus(directory)
         log_file = path.join(directory, "irida-uploader.log")
-
         # Check that log file does not exist before starting
         self.assertFalse(path.exists(log_file))
 
-        cli_entry.validate_and_upload_single_entry(directory)
+        cli_entry._validate_and_upload(directory_status)
 
         # Make sure log file is created
         self.assertTrue(path.exists(log_file))
@@ -110,6 +112,9 @@ class TestValidateAndUploadSingleEntry(unittest.TestCase):
         When parsing, it should be valid
         :return:
         """
+
+        directory = path.join(path_to_module, "fake_ngs_data")
+
         class StubValidationResult:
             @staticmethod
             def is_valid():
@@ -132,15 +137,14 @@ class TestValidateAndUploadSingleEntry(unittest.TestCase):
                 return status == DirectoryStatus.NEW
 
         # add mock data to the function calls that are essential
-        mock_parsing_handler.get_run_status.side_effect = [StubDirectoryStatus]
+        stub_directory_status = StubDirectoryStatus()
+        mock_parsing_handler.get_run_status.side_effect = [stub_directory_status]
         mock_parsing_handler.parse_and_validate.side_effect = ["Fake Sequencing Run"]
         mock_api_handler.initialize_api_from_config.side_effect = [None]
         mock_api_handler.prepare_and_validate_for_upload.side_effect = [StubValidationResult]
         mock_api_handler.upload_sequencing_run.side_effect = [None]
 
-        directory = path.join(path_to_module, "fake_ngs_data")
-
-        cli_entry.validate_and_upload_single_entry(directory)
+        cli_entry.upload_run_single_entry(directory)
 
         # Make sure the validation is tried
         mock_api_handler.prepare_and_validate_for_upload.assert_called_with("Fake Sequencing Run")
@@ -174,7 +178,7 @@ class TestValidateAndUploadSingleEntry(unittest.TestCase):
 
         directory = path.join(path_to_module, "fake_ngs_data")
 
-        cli_entry.validate_and_upload_single_entry(directory)
+        cli_entry.upload_run_single_entry(directory)
 
         # Check that run status was found
         mock_parsing_handler.get_run_status.assert_called_with(directory)
@@ -193,6 +197,7 @@ class TestValidateAndUploadSingleEntry(unittest.TestCase):
         Invalidity comes from parsing module
         :return:
         """
+        directory = path.join(path_to_module, "fake_ngs_data")
 
         class StubDirectoryStatus:
             directory = path.join(path_to_module, "fake_ngs_data")
@@ -209,9 +214,8 @@ class TestValidateAndUploadSingleEntry(unittest.TestCase):
         mock_api_handler.prepare_and_validate_for_upload.side_effect = [None]
         mock_api_handler.upload_sequencing_run.side_effect = [None]
 
-        directory = path.join(path_to_module, "fake_ngs_data")
 
-        cli_entry.validate_and_upload_single_entry(directory)
+        cli_entry.upload_run_single_entry(directory)
 
         # Check that run status was found
         mock_parsing_handler.get_run_status.assert_called_with(directory)
@@ -231,6 +235,9 @@ class TestValidateAndUploadSingleEntry(unittest.TestCase):
         In this case, we are forcing an upload, so no run is new check should happen
         :return:
         """
+
+        directory = path.join(path_to_module, "fake_ngs_data")
+
         class StubValidationResult:
             @staticmethod
             def is_valid():
@@ -250,9 +257,7 @@ class TestValidateAndUploadSingleEntry(unittest.TestCase):
         mock_api_handler.prepare_and_validate_for_upload.side_effect = [StubValidationResult]
         mock_api_handler.upload_sequencing_run.side_effect = [None]
 
-        directory = path.join(path_to_module, "fake_ngs_data")
-
-        cli_entry.validate_and_upload_single_entry(directory, True)
+        cli_entry.upload_run_single_entry(directory, True)
 
         # Check that run status was found
         mock_parsing_handler.get_run_status.assert_called_with(directory)
@@ -276,6 +281,8 @@ class TestValidateAndUploadSingleEntry(unittest.TestCase):
         In this case, the run is new check should happen
         :return:
         """
+        directory = path.join(path_to_module, "fake_ngs_data")
+
         class StubValidationResult:
             @staticmethod
             def is_valid():
@@ -295,9 +302,7 @@ class TestValidateAndUploadSingleEntry(unittest.TestCase):
         mock_api_handler.prepare_and_validate_for_upload.side_effect = [StubValidationResult]
         mock_api_handler.upload_sequencing_run.side_effect = [None]
 
-        directory = path.join(path_to_module, "fake_ngs_data")
-
-        cli_entry.validate_and_upload_single_entry(directory, False)
+        cli_entry.upload_run_single_entry(directory, False)
 
         # Check that run status was found
         mock_parsing_handler.get_run_status.assert_called_with(directory)
@@ -321,6 +326,9 @@ class TestValidateAndUploadSingleEntry(unittest.TestCase):
         In this case, the run is new check should occur, and the run should not be uploaded, and not be parsed
         :return:
         """
+
+        directory = path.join(path_to_module, "fake_ngs_data")
+
         class StubValidationResult:
             @staticmethod
             def is_valid():
@@ -341,9 +349,7 @@ class TestValidateAndUploadSingleEntry(unittest.TestCase):
         mock_api_handler.prepare_and_validate_for_upload.side_effect = [StubValidationResult]
         mock_api_handler.upload_sequencing_run.side_effect = [None]
 
-        directory = path.join(path_to_module, "fake_ngs_data")
-
-        cli_entry.validate_and_upload_single_entry(directory, False)
+        cli_entry.upload_run_single_entry(directory, False)
 
         # Check that run status was found
         mock_parsing_handler.get_run_status.assert_called_with(directory)
@@ -353,3 +359,91 @@ class TestValidateAndUploadSingleEntry(unittest.TestCase):
         mock_parsing_handler.parse_and_validate.assert_not_called()
         # make sure the upload is NOT done, as validation is invalid
         mock_api_handler.upload_sequencing_run.assert_not_called()
+
+
+class TestBatchUploadSingleEntry(unittest.TestCase):
+    """
+    Tests the core.cli_entry.batch_upload_single_entry function
+    """
+
+    def setUp(self):
+        print("\nStarting " + self.__module__ + ": " + self._testMethodName)
+
+    @patch("core.cli_entry._validate_and_upload")
+    @patch("core.cli_entry.parsing_handler")
+    def test_valid(self, mock_parsing_handler, mock_validate_and_upload):
+        """
+        Makes sure that _validate_and_upload is only called on a new run
+        :return:
+        """
+        class StubDirectoryStatus:
+            def __init__(self, directory, status):
+                self.directory = directory
+                self.status = status
+                self.message = None
+
+            def status_equals(self, other_status):
+                return self.status == other_status
+
+        # add mock data to the function calls that are essential
+        stub_directory_status_valid = StubDirectoryStatus("valid", DirectoryStatus.NEW)
+        stub_directory_status_invalid = StubDirectoryStatus("invalid", DirectoryStatus.INVALID)
+        stub_directory_status_complete = StubDirectoryStatus("complete", DirectoryStatus.COMPLETE)
+        stub_directory_status_partial = StubDirectoryStatus("partial", DirectoryStatus.PARTIAL)
+
+        mock_parsing_handler.get_run_status_list.side_effect = [
+            [stub_directory_status_valid,
+             stub_directory_status_invalid,
+             stub_directory_status_complete,
+             stub_directory_status_partial]
+        ]
+        mock_parsing_handler.parse_and_validate.side_effect = ["Fake Sequencing Run"]
+
+        # start
+        cli_entry.batch_upload_single_entry("fake_directory", force_upload=False)
+
+        # validate calls only happen once
+        mock_validate_and_upload.assert_called_once_with(stub_directory_status_valid)
+
+
+    @patch("core.cli_entry._validate_and_upload")
+    @patch("core.cli_entry.parsing_handler")
+    def test_valid_force(self, mock_parsing_handler, mock_validate_and_upload):
+        """
+        Makes sure that _validate_and_upload is called on all runs except invalid
+        :return:
+        """
+        class StubDirectoryStatus:
+            def __init__(self, directory, status):
+                self.directory = directory
+                self.status = status
+                self.message = None
+
+            def status_equals(self, other_status):
+                return self.status == other_status
+
+        # add mock data to the function calls that are essential
+        stub_directory_status_valid = StubDirectoryStatus("valid", DirectoryStatus.NEW)
+        stub_directory_status_invalid = StubDirectoryStatus("invalid", DirectoryStatus.INVALID)
+        stub_directory_status_complete = StubDirectoryStatus("complete", DirectoryStatus.COMPLETE)
+        stub_directory_status_partial = StubDirectoryStatus("partial", DirectoryStatus.PARTIAL)
+
+        mock_parsing_handler.get_run_status_list.side_effect = [
+            [stub_directory_status_valid,
+             stub_directory_status_invalid,
+             stub_directory_status_complete,
+             stub_directory_status_partial]
+        ]
+        mock_parsing_handler.parse_and_validate.side_effect = ["Fake Sequencing Run"]
+
+        # start
+        cli_entry.batch_upload_single_entry("fake_directory", force_upload=True)
+
+        # assert calls are what we expect
+        self.assertEqual(mock_validate_and_upload.call_count, 3, "Expected 3 calls to mock_validate_and_upload")
+        expected_call_args = [
+            call(stub_directory_status_valid,),
+            call(stub_directory_status_complete,),
+            call(stub_directory_status_partial,)
+        ]
+        self.assertEqual(mock_validate_and_upload.call_args_list, expected_call_args, "Call args do not match expected")