+++ tests_integration/test_end_to_end.py
@@ -4,13 +4,12 @@ import os
 
 from tests_integration import tests_integration
 
-import global_settings
 import config
 import api
 import model
 import progress
 
-from core.cli_entry import validate_and_upload_single_entry
+from core.cli_entry import upload_run_single_entry, batch_upload_single_entry
 
 
 path_to_module = path.dirname(__file__)
@@ -20,11 +19,15 @@ if len(path_to_module) == 0:
 CLEANUP_DIRECTORY_LIST = [
     path.join(path_to_module, "fake_dir_data"),
     path.join(path_to_module, "fake_miniseq_data"),
+    path.join(path_to_module, "fake_nextseq_data"),
     path.join(path_to_module, "fake_ngs_data"),
     path.join(path_to_module, "fake_ngs_data_force"),
     path.join(path_to_module, "fake_ngs_data_nonexistent_project"),
     path.join(path_to_module, "fake_ngs_data_parse_fail"),
-    path.join(path_to_module, "fake_ngs_data_no_completed_file")
+    path.join(path_to_module, "fake_ngs_data_no_completed_file"),
+    path.join(path_to_module, "fake_batch_data", "run_1"),
+    path.join(path_to_module, "fake_batch_data", "run_2"),
+    path.join(path_to_module, "fake_batch_data", "run_3")
 ]
 
 
@@ -38,7 +41,7 @@ class TestEndToEnd(unittest.TestCase):
     def setUp(self):
         print("\nStarting " + self.__module__ + ": " + self._testMethodName)
         # Set the config file to our test config, so that when the upload is run it grabs out config file correctly
-        global_settings.config_file = path.join(path_to_module, "test_config.conf")
+        config.set_config_file(path.join(path_to_module, "test_config.conf"))
         config.setup()
 
     def tearDown(self):
@@ -70,12 +73,13 @@ class TestEndToEnd(unittest.TestCase):
         :param parser:
         :return:
         """
-        config.write_config_option("client_id", client_id)
-        config.write_config_option("client_secret", client_secret)
-        config.write_config_option("username", username)
-        config.write_config_option("password", password)
-        config.write_config_option("base_url", base_url)
-        config.write_config_option("parser", parser)
+        config.set_config_options(client_id=client_id,
+                                  client_secret=client_secret,
+                                  username=username,
+                                  password=password,
+                                  base_url=base_url,
+                                  parser=parser)
+        config.write_config_options_to_file()
 
     def test_valid_miseq_upload(self):
         """
@@ -113,7 +117,7 @@ class TestEndToEnd(unittest.TestCase):
         project_id = "1"
 
         # Do the upload
-        upload_result = validate_and_upload_single_entry(path.join(path_to_module, "fake_ngs_data"))
+        upload_result = upload_run_single_entry(path.join(path_to_module, "fake_ngs_data"))
 
         # Make sure the upload was a success
         self.assertEqual(upload_result, 0)
@@ -171,6 +175,95 @@ class TestEndToEnd(unittest.TestCase):
         self.assertEqual(sample_2_found, True)
         self.assertEqual(sample_3_found, True)
 
+    def test_batch_miseq_upload(self):
+        """
+        Test a valid miseq directory for upload from end to end
+        We have 3 run directories
+            run_1 has batch01-1111
+            run_2 is invalid
+            run_3 has batch03-3333
+        we expect to see batch01-1111 and batch03-3333 uploaded
+        :return:
+        """
+        # Set our sample config file to use miseq parser and the correct irida credentials
+        self.write_to_config_file(
+            client_id=tests_integration.client_id,
+            client_secret=tests_integration.client_secret,
+            username=tests_integration.username,
+            password=tests_integration.password,
+            base_url=tests_integration.base_url,
+            parser="miseq"
+        )
+
+        # instance an api
+        test_api = api.ApiCalls(
+            client_id=tests_integration.client_id,
+            client_secret=tests_integration.client_secret,
+            base_url=tests_integration.base_url,
+            username=tests_integration.username,
+            password=tests_integration.password
+        )
+
+        # Create a test project, the uploader does not make new projects on its own
+        # so one must exist to upload samples into
+        # This may not be the project that the files get uploaded to,
+        # but one will be made in the case this is the only test being run
+        project_name = "test_batch_project"
+        project_description = "test_batch_project_description"
+        project = model.Project(name=project_name, description=project_description)
+        test_api.send_project(project)
+        # We always upload to project "1" so that tests will be consistent no matter how many / which tests are run
+        project_id = "1"
+
+        # Do the upload
+        upload_result = batch_upload_single_entry(path.join(path_to_module, "fake_batch_data"))
+
+        # Make sure the upload was a success
+        self.assertEqual(upload_result, 0)
+
+        # Verify the files were uploaded
+        sample_list = test_api.get_samples(project_id)
+
+        sample_1_found = False
+        sample_2_not_found = True
+        sample_3_found = False
+
+        for sample in sample_list:
+            if sample.sample_name in ["batch01-1111", "batch02-2222", "batch03-3333"]:
+                if sample.sample_name == "batch01-1111":
+                    sample_1_found = True
+                    sequence_files = test_api.get_sequence_files(project_id, sample.sample_name)
+                    self.assertEqual(len(sequence_files), 2)
+                    res_sequence_file_names = [
+                        sequence_files[0]['fileName'],
+                        sequence_files[1]['fileName']
+                    ]
+                    expected_sequence_file_names = [
+                        'batch01-1111_S1_L001_R1_001.fastq.gz',
+                        'batch01-1111_S1_L001_R2_001.fastq.gz'
+                    ]
+                    self.assertEqual(res_sequence_file_names.sort(), expected_sequence_file_names.sort())
+                elif sample.sample_name == "batch02-2222":
+                    # this one should not be found
+                    sample_2_not_found = False
+                elif sample.sample_name == "batch03-3333":
+                    sample_3_found = True
+                    sequence_files = test_api.get_sequence_files(project_id, sample.sample_name)
+                    self.assertEqual(len(sequence_files), 2)
+                    res_sequence_file_names = [
+                        sequence_files[0]['fileName'],
+                        sequence_files[1]['fileName']
+                    ]
+                    expected_sequence_file_names = [
+                        'batch03-3333_S1_L001_R1_001.fastq.gz',
+                        'batch03-3333_S1_L001_R2_001.fastq.gz'
+                    ]
+                    self.assertEqual(res_sequence_file_names.sort(), expected_sequence_file_names.sort())
+
+        self.assertEqual(sample_1_found, True)
+        self.assertEqual(sample_2_not_found, True)
+        self.assertEqual(sample_3_found, True)
+
     def test_valid_directory_upload(self):
         """
         Test a valid directory for upload end to end
@@ -207,7 +300,7 @@ class TestEndToEnd(unittest.TestCase):
         project_id = "1"
 
         # Do the upload
-        upload_result = validate_and_upload_single_entry(path.join(path_to_module, "fake_dir_data"))
+        upload_result = upload_run_single_entry(path.join(path_to_module, "fake_dir_data"))
 
         # Make sure the upload was a success
         self.assertEqual(upload_result, 0)
@@ -280,7 +373,7 @@ class TestEndToEnd(unittest.TestCase):
         project_id = "1"
 
         # Do the upload
-        upload_result = validate_and_upload_single_entry(path.join(path_to_module, "fake_miniseq_data"))
+        upload_result = upload_run_single_entry(path.join(path_to_module, "fake_miniseq_data"))
 
         # Make sure the upload was a success
         self.assertEqual(upload_result, 0)
@@ -338,6 +431,94 @@ class TestEndToEnd(unittest.TestCase):
         self.assertEqual(sample_2_found, True)
         self.assertEqual(sample_3_found, True)
 
+    def test_valid_nextseq_upload(self):
+        """
+        Test a valid nextseq directory for upload from end to end
+        :return:
+        """
+        # Set our sample config file to use miseq parser and the correct irida credentials
+        self.write_to_config_file(
+            client_id=tests_integration.client_id,
+            client_secret=tests_integration.client_secret,
+            username=tests_integration.username,
+            password=tests_integration.password,
+            base_url=tests_integration.base_url,
+            parser="nextseq"
+        )
+
+        # instance an api
+        test_api = api.ApiCalls(
+            client_id=tests_integration.client_id,
+            client_secret=tests_integration.client_secret,
+            base_url=tests_integration.base_url,
+            username=tests_integration.username,
+            password=tests_integration.password
+        )
+
+        # Create a test project, the uploader does not make new projects on its own
+        # so one must exist to upload samples into
+        # This may not be the project that the files get uploaded to,
+        # but one will be made in the case this is the only test being run
+        project_name = "test_project_nextseq"
+        project_description = "test_project_description_nextseq"
+        project = model.Project(name=project_name, description=project_description)
+        test_api.send_project(project)
+        # We always upload to project "1" so that tests will be consistent no matter how many / which tests are run
+        project_id_1 = "1"
+
+        # we are uploading 2 projects, so create another one
+        project_name_2 = "test_project_nextseq_2"
+        project_description_2 = "test_project_description_nextseq_2"
+        project_2 = model.Project(name=project_name_2, description=project_description_2)
+        test_api.send_project(project_2)
+        project_id_2 = "2"
+
+        # Do the upload
+        upload_result = upload_run_single_entry(path.join(path_to_module, "fake_nextseq_data"))
+
+        # Make sure the upload was a success
+        self.assertEqual(upload_result, 0)
+
+        # Verify the files were uploaded
+        sample_list_1 = test_api.get_samples(project_id_1)
+        sample_list_2 = test_api.get_samples(project_id_2)
+
+        sample_1_found = False
+        sample_2_found = False
+
+        for sample in sample_list_1:
+            if sample.sample_name == "SA20121712":
+                sample_1_found = True
+                sequence_files = test_api.get_sequence_files(project_id_1, sample.sample_name)
+                self.assertEqual(len(sequence_files), 2)
+                res_sequence_file_names = [
+                    sequence_files[0]['fileName'],
+                    sequence_files[1]['fileName']
+                ]
+                expected_sequence_file_names = [
+                    'SA20121712_S2_R1_001.fastq.qz',
+                    'SA20121712_S2_R2_001.fastq.qz'
+                ]
+                self.assertEqual(res_sequence_file_names.sort(), expected_sequence_file_names.sort())
+
+        for sample in sample_list_2:
+            if sample.sample_name == "SA20121716":
+                sample_2_found = True
+                sequence_files = test_api.get_sequence_files(project_id_2, sample.sample_name)
+                self.assertEqual(len(sequence_files), 2)
+                res_sequence_file_names = [
+                    sequence_files[0]['fileName'],
+                    sequence_files[1]['fileName']
+                ]
+                expected_sequence_file_names = [
+                    'SA20121716_S1_R1_001.fastq.qz',
+                    'SA20121716_S1_R2_001.fastq.qz'
+                ]
+                self.assertEqual(res_sequence_file_names.sort(), expected_sequence_file_names.sort())
+
+        self.assertEqual(sample_1_found, True)
+        self.assertEqual(sample_2_found, True)
+
     def test_upload_to_nonexistent_project(self):
         """
         Everything is correct except the sample sheet file specifies an invalid project
@@ -374,7 +555,7 @@ class TestEndToEnd(unittest.TestCase):
         test_api.send_project(project)
 
         # Do the upload
-        upload_result = validate_and_upload_single_entry(path.join(path_to_module, "fake_ngs_data_nonexistent_project"))
+        upload_result = upload_run_single_entry(path.join(path_to_module, "fake_ngs_data_nonexistent_project"))
 
         # Make sure the upload was a failure
         self.assertEqual(upload_result, 1)
@@ -419,7 +600,7 @@ class TestEndToEnd(unittest.TestCase):
         test_api.send_project(project)
 
         # Do the upload
-        upload_result = validate_and_upload_single_entry(path.join(path_to_module, "fake_ngs_data_parse_fail"))
+        upload_result = upload_run_single_entry(path.join(path_to_module, "fake_ngs_data_parse_fail"))
 
         # Make sure the upload was a failure
         self.assertEqual(upload_result, 1)
@@ -467,7 +648,7 @@ class TestEndToEnd(unittest.TestCase):
         progress.write_directory_status(directory_status)
 
         # Do the upload, with force option
-        upload_result = validate_and_upload_single_entry(path.join(path_to_module, "fake_ngs_data_force"), True)
+        upload_result = upload_run_single_entry(path.join(path_to_module, "fake_ngs_data_force"), True)
 
         # Make sure the upload was a success
         self.assertEqual(upload_result, 0)
@@ -548,7 +729,7 @@ class TestEndToEnd(unittest.TestCase):
         progress.write_directory_status(directory_status)
 
         # Do the upload, without force option
-        upload_result = validate_and_upload_single_entry(path.join(path_to_module, "fake_ngs_data"), False)
+        upload_result = upload_run_single_entry(path.join(path_to_module, "fake_ngs_data"), False)
 
         # Make sure the upload was a failure
         self.assertEqual(upload_result, 1)
@@ -571,7 +752,7 @@ class TestEndToEnd(unittest.TestCase):
         )
 
         # Do the upload, without force option
-        upload_result = validate_and_upload_single_entry(path.join(path_to_module, "fake_ngs_data_no_completed_file"), False)
+        upload_result = upload_run_single_entry(path.join(path_to_module, "fake_ngs_data_no_completed_file"), False)
 
         # Make sure the upload was a failure
         self.assertEqual(upload_result, 1)