+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 Changes
 =======
 
+Beta 0.3
+--------
+Added functionality:
+* Added config option overrides for the command line
+    * Can define in command line `--option parameter` or without a parameter `--option` for a user prompt
+    * Use the `--help` option for more details on these options
+
+Beta 0.2.1
+----------
+Bug Fixes:
+* Fixed bug where 2 description fields would be sent with some miniseq runs, causing upload to fail.
 Beta 0.2
 --------
 Added functionality:
@@ -12,7 +23,9 @@ Added functionality:
 * When uploading, a progress percentage will be printed to the command line (Linux + Windows)
 * Miseq uploader now requires `CompletedJobInfo.xml` file to upload
 * Added support for MiniSeq, use parser `miniseq` to use.
+* Added support for NextSeq, use parser `nextseq` to use.
 * Added shebang to top of upload_run.py to support conda environments
+* Added batch uploading, use `--batch` or `-b` with a directory containing multiple runs to upload.
 
 Beta 0.1
 ---------